draft join (no tests)
