+++ src/dex/uniswap-v2/constants.ts
@@ -4,8 +4,7 @@ import { MDEXConfig } from './mdex';
 import { BiSwapConfig } from './biswap';
 import { DfynConfig } from './dfyn';
 import { ExcaliburConfig } from './excalibur';
-import { DystopiaConfig } from './dystopia/dystopia-volatile';
-import { DystopiaStableConfig } from './dystopia/dystopia-stable';
+import { DystopiaConfig } from './dystopia/dystopia';
 
 // BakerySwap is removed from AllUniswapForks and UniswapForksWithNetwork
 // as it has a modified pool implementation which is not compatible with
@@ -18,7 +17,6 @@ export const AllUniswapForks = [
   ...Object.keys(DfynConfig),
   ...Object.keys(ExcaliburConfig),
   ...Object.keys(DystopiaConfig),
-  ...Object.keys(DystopiaStableConfig),
 ];
 
 const transformToNetworkMap = (config: {
@@ -47,7 +45,6 @@ export const UniswapForksWithNetwork = transformToNetworkMap({
   ...DfynConfig,
   ...ExcaliburConfig,
   ...DystopiaConfig,
-  ...DystopiaStableConfig,
 });
 
 // These are exchanges used for swapOnUniswap method
