+++ src/dex/uniswap-v2/dystopia/dystopia-constants.ts
@@ -1,0 +1,2 @@
+/// 0.05% swap fee
+export const SWAP_FEE_FACTOR = 2000n;
