+++ src/dex/uniswap-v2/dystopia/dystopia.ts
@@ -1,4 +1,4 @@
-import { RESERVE_LIMIT, UniswapV2 } from '../uniswap-v2';
+import { RESERVE_LIMIT, UniswapV2, UniswapV2Pair } from '../uniswap-v2';
 import { Network, NULL_ADDRESS, subgraphTimeout } from '../../../constants';
 import {
   AdapterExchangeParam,
@@ -21,6 +21,7 @@ import dystopiaFactoryABI from '../../../abi/uniswap-v2/DystFactory.json';
 import { BI_MAX_UINT } from '../../../bigint-constants';
 import _ from 'lodash';
 import { NumberAsString, SwapSide } from 'paraswap-core';
+import { SWAP_FEE_FACTOR } from './dystopia-constants';
 
 export const DystopiaConfig: DexConfigMap<DexParams> = {
   Dystopia: {
@@ -38,10 +39,21 @@ export const DystopiaConfig: DexConfigMap<DexParams> = {
   },
 };
 
-export class Dystopia extends UniswapV2 {
-  /// 0.05% swap fee
-  private static SWAP_FEE_FACTOR: 2000n;
+// to calculate prices for stable pool, we need decimals of the stable tokens
+// so, we are extending UniswapV2PoolOrderedParams with token decimals
+export interface UniswapV2PoolOrderedParamsWithDecimals {
+  tokenIn: string;
+  tokenOut: string;
+  reservesIn: string;
+  reservesOut: string;
+  fee: string;
+  direction: boolean;
+  exchange: string;
+  decimalsIn: number;
+  decimalsOut: number;
+}
 
+export class Dystopia extends UniswapV2 {
   public static dexKeysWithNetwork: { key: string; networks: Network[] }[] =
     getDexKeysWithNetwork(DystopiaConfig);
 
@@ -68,22 +80,22 @@ export class Dystopia extends UniswapV2 {
     );
   }
 
-  async findPair(from: Token, to: Token) {
+  async findDystopiaPair(from: Token, to: Token, stable: boolean) {
     if (from.address.toLowerCase() === to.address.toLowerCase()) return null;
     const [token0, token1] =
       from.address.toLowerCase() < to.address.toLowerCase()
         ? [from, to]
         : [to, from];
 
-    const key = `${token0.address.toLowerCase()}-${token1.address.toLowerCase()}`;
+    const typePostfix = this.poolPostfix(stable);
+    const key = `${token0.address.toLowerCase()}-${token1.address.toLowerCase()}-${typePostfix}`;
     let pair = this.pairs[key];
     if (pair) return pair;
 
     let exchange = await this.factory.methods
       // Dystopia has additional boolean parameter "StablePool"
-      // At this Dystopia implementation we're looking for
-      // non-stable (uniswap2-like) pools only
-      .getPair(token0.address, token1.address, false)
+      // At first we look for uniswap-like volatile pool
+      .getPair(token0.address, token1.address, stable)
       .call();
 
     if (exchange === NULL_ADDRESS) {
@@ -93,6 +105,46 @@ export class Dystopia extends UniswapV2 {
     }
     this.pairs[key] = pair;
     return pair;
+  }
+
+  async batchCatchUpPairs(pairs: [Token, Token][], blockNumber: number) {
+    if (!blockNumber) return;
+    const pairsToFetch: UniswapV2Pair[] = [];
+    for (const _pair of pairs) {
+      for (const stable of [false, true]) {
+        const pair = await this.findDystopiaPair(_pair[0], _pair[1], stable);
+        if (!(pair && pair.exchange)) continue;
+        if (!pair.pool) {
+          pairsToFetch.push(pair);
+        } else if (!pair.pool.getState(blockNumber)) {
+          pairsToFetch.push(pair);
+        }
+      }
+    }
+
+    if (!pairsToFetch.length) return;
+
+    const reserves = await this.getManyPoolReserves(pairsToFetch, blockNumber);
+
+    if (reserves.length !== pairsToFetch.length) {
+      this.logger.error(
+        `Error_getManyPoolReserves didn't get any pool reserves`,
+      );
+    }
+
+    for (let i = 0; i < pairsToFetch.length; i++) {
+      const pairState = reserves[i];
+      const pair = pairsToFetch[i];
+      if (!pair.pool) {
+        await this.addPool(
+          pair,
+          pairState.reserves0,
+          pairState.reserves1,
+          pairState.feeCode,
+          blockNumber,
+        );
+      } else pair.pool.setState(pairState, blockNumber);
+    }
   }
 
   // Dystopia non-stable pools has almost same formula like uniswap2,
@@ -108,12 +160,11 @@ export class Dystopia extends UniswapV2 {
       return 0n;
     }
 
-    const amountInWithFee = srcAmount - srcAmount / Dystopia.SWAP_FEE_FACTOR;
+    const amountInWithFee = srcAmount - srcAmount / SWAP_FEE_FACTOR;
 
     const numerator = amountInWithFee * BigInt(reservesOut);
 
-    const denominator =
-      BigInt(reservesIn) * Dystopia.SWAP_FEE_FACTOR + amountInWithFee;
+    const denominator = BigInt(reservesIn) * SWAP_FEE_FACTOR + amountInWithFee;
 
     return denominator === 0n ? 0n : numerator / denominator;
   }
@@ -124,10 +175,9 @@ export class Dystopia extends UniswapV2 {
   ): Promise<bigint> {
     const { reservesIn, reservesOut } = priceParams;
 
-    const numerator =
-      BigInt(reservesIn) * destAmount * Dystopia.SWAP_FEE_FACTOR;
+    const numerator = BigInt(reservesIn) * destAmount * SWAP_FEE_FACTOR;
     const denominator =
-      (Dystopia.SWAP_FEE_FACTOR - 1n) * (BigInt(reservesOut) - destAmount);
+      (SWAP_FEE_FACTOR - 1n) * (BigInt(reservesOut) - destAmount);
 
     if (denominator <= 0n) return BI_MAX_UINT;
     return 1n + numerator / denominator;
@@ -180,6 +230,7 @@ export class Dystopia extends UniswapV2 {
       throw new Error("Couldn't fetch the pools from the subgraph");
     const pools0 = _.map(data.pools0, pool => ({
       exchange: this.dexKey,
+      stable: pool.isStable,
       address: pool.id.toLowerCase(),
       connectorTokens: [
         {
@@ -192,6 +243,7 @@ export class Dystopia extends UniswapV2 {
 
     const pools1 = _.map(data.pools1, pool => ({
       exchange: this.dexKey,
+      stable: pool.isStable,
       address: pool.id.toLowerCase(),
       connectorTokens: [
         {
@@ -207,6 +259,22 @@ export class Dystopia extends UniswapV2 {
       0,
       count,
     );
+  }
+
+  // Same as at uniswap-v2-pool.json, but extended with decimals
+  async getPairOrderedParams(
+    from: Token,
+    to: Token,
+    blockNumber: number,
+  ): Promise<UniswapV2PoolOrderedParamsWithDecimals | null> {
+    const params = await super.getPairOrderedParams(from, to, blockNumber);
+    if (!params) return null;
+
+    return {
+      ...params,
+      decimalsIn: from.decimals,
+      decimalsOut: to.decimals,
+    };
   }
 
   async getPricesVolume(
@@ -248,9 +316,13 @@ export class Dystopia extends UniswapV2 {
       .join('_');
 
     const poolIdentifier = `${this.dexKey}_${tokenAddress}`;
-    const poolIdentifierUniswap = poolIdentifier + '_u';
-    const poolIdentifierStable = poolIdentifier + '_s';
+    const poolIdentifierUniswap = poolIdentifier + this.poolPostfix(false);
+    const poolIdentifierStable = poolIdentifier + this.poolPostfix(true);
     return [poolIdentifierUniswap, poolIdentifierStable];
+  }
+
+  poolPostfix(stable: boolean) {
+    return stable ? 'S' : 'U';
   }
 
   async getSimpleParam(
