+++ src/dex/uniswap-v2/uniswap-v2-e2e-polygon.test.ts
@@ -1238,192 +1238,191 @@ describe('UniswapV2 E2E Polygon', () => {
 
   describe('Dystopia', () => {
     const dexKey = 'Dystopia';
-
-    const maticAmount = '1000000000000000000';
     const usdAmount = '1000000';
 
-    describe('simpleSwap', () => {
-      it('MATIC -> TOKEN', async () => {
-        await testE2E(
-          tokens.MATIC,
-          tokens.WETH,
-          holders.MATIC,
-          maticAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.simpleSwap,
-          network,
-          provider,
-        );
+    describe('Dystopia UniswapV2 Pools', () => {
+      const maticAmount = '1000000000000000000';
+
+      describe('simpleSwap', () => {
+        it('MATIC -> TOKEN', async () => {
+          await testE2E(
+            tokens.MATIC,
+            tokens.WETH,
+            holders.MATIC,
+            maticAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.simpleSwap,
+            network,
+            provider,
+          );
+        });
+
+        it('Token -> MATIC', async () => {
+          await testE2E(
+            tokens.USDT,
+            tokens.MATIC,
+            holders.USDT,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.simpleSwap,
+            network,
+            provider,
+          );
+        });
+
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.WMATIC,
+            tokens.USDT,
+            holders.WMATIC,
+            maticAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.simpleSwap,
+            network,
+            provider,
+          );
+        });
       });
 
-      it('Token -> MATIC', async () => {
-        await testE2E(
-          tokens.USDT,
-          tokens.MATIC,
-          holders.USDT,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.simpleSwap,
-          network,
-          provider,
-        );
+      describe('multiSwap', () => {
+        it('MATIC -> TOKEN', async () => {
+          await testE2E(
+            tokens.MATIC,
+            tokens.WETH,
+            holders.MATIC,
+            maticAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.multiSwap,
+            network,
+            provider,
+          );
+        });
+
+        it('Token -> MATIC', async () => {
+          await testE2E(
+            tokens.USDT,
+            tokens.MATIC,
+            holders.USDT,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.multiSwap,
+            network,
+            provider,
+          );
+        });
+
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.WMATIC,
+            tokens.USDT,
+            holders.WMATIC,
+            maticAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.multiSwap,
+            network,
+            provider,
+          );
+        });
       });
 
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.WMATIC,
-          tokens.USDT,
-          holders.WMATIC,
-          maticAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.simpleSwap,
-          network,
-          provider,
-        );
-      });
-    });
+      describe('megaSwap', () => {
+        it('MATIC -> TOKEN', async () => {
+          await testE2E(
+            tokens.USDT,
+            tokens.MATIC,
+            holders.USDT,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.megaSwap,
+            network,
+            provider,
+          );
+        });
 
-    describe('multiSwap', () => {
-      it('MATIC -> TOKEN', async () => {
-        await testE2E(
-          tokens.MATIC,
-          tokens.WETH,
-          holders.MATIC,
-          maticAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.multiSwap,
-          network,
-          provider,
-        );
-      });
+        it('Token -> MATIC', async () => {
+          await testE2E(
+            tokens.USDT,
+            tokens.MATIC,
+            holders.USDT,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.megaSwap,
+            network,
+            provider,
+          );
+        });
 
-      it('Token -> MATIC', async () => {
-        await testE2E(
-          tokens.USDT,
-          tokens.MATIC,
-          holders.USDT,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.multiSwap,
-          network,
-          provider,
-        );
-      });
-
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.WMATIC,
-          tokens.USDT,
-          holders.WMATIC,
-          maticAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.multiSwap,
-          network,
-          provider,
-        );
-      });
-    });
-
-    describe('megaSwap', () => {
-      it('MATIC -> TOKEN', async () => {
-        await testE2E(
-          tokens.USDT,
-          tokens.MATIC,
-          holders.USDT,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.megaSwap,
-          network,
-          provider,
-        );
-      });
-
-      it('Token -> MATIC', async () => {
-        await testE2E(
-          tokens.USDT,
-          tokens.MATIC,
-          holders.USDT,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.megaSwap,
-          network,
-          provider,
-        );
-      });
-
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.WMATIC,
-          tokens.USDT,
-          holders.WMATIC,
-          maticAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.megaSwap,
-          network,
-          provider,
-        );
-      });
-    });
-  });
-
-  describe('DystopiaStable', () => {
-    const dexKey = 'DystopiaStable';
-
-    const usdAmount = '1000000';
-
-    describe('simpleSwap', () => {
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.USDC,
-          tokens.USDT,
-          holders.USDC,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.simpleSwap,
-          network,
-          provider,
-        );
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.WMATIC,
+            tokens.USDT,
+            holders.WMATIC,
+            maticAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.megaSwap,
+            network,
+            provider,
+          );
+        });
       });
     });
 
-    describe('multiSwap', () => {
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.USDC,
-          tokens.USDT,
-          holders.USDC,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.multiSwap,
-          network,
-          provider,
-        );
+    describe('Dystopia Stable Pools', () => {
+      describe('simpleSwap', () => {
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.USDC,
+            tokens.USDT,
+            holders.USDC,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.simpleSwap,
+            network,
+            provider,
+          );
+        });
       });
-    });
-    describe('megaSwap', () => {
-      it('Token -> Token', async () => {
-        await testE2E(
-          tokens.USDC,
-          tokens.USDT,
-          holders.USDC,
-          usdAmount,
-          SwapSide.SELL,
-          dexKey,
-          ContractMethod.megaSwap,
-          network,
-          provider,
-        );
+
+      describe('multiSwap', () => {
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.USDC,
+            tokens.USDT,
+            holders.USDC,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.multiSwap,
+            network,
+            provider,
+          );
+        });
+      });
+
+      describe('megaSwap', () => {
+        it('Token -> Token', async () => {
+          await testE2E(
+            tokens.USDC,
+            tokens.USDT,
+            holders.USDC,
+            usdAmount,
+            SwapSide.SELL,
+            dexKey,
+            ContractMethod.megaSwap,
+            network,
+            provider,
+          );
+        });
       });
     });
   });
