+++ src/dex/uniswap-v2/uniswap-v2-integration.test.ts
@@ -6,8 +6,7 @@ import { Network, SwapSide } from '../../constants';
 import { UniswapV2 } from './uniswap-v2';
 import { checkPoolPrices, checkPoolsLiquidity } from '../../../tests/utils';
 import { BI_POWS } from '../../bigint-constants';
-import { DystopiaVolatile } from './dystopia/dystopia-volatile';
-import { DystopiaStable } from './dystopia/dystopia-stable';
+import { Dystopia } from './dystopia/dystopia';
 import { Tokens } from '../../../tests/constants-e2e';
 
 const WETH = {
@@ -66,8 +65,8 @@ describe('UniswapV2', function () {
 });
 
 describe('Dystopia', function () {
+  const dexKey = 'Dystopia';
   describe('UniswapV2 like pool', function () {
-    const dexKey = 'Dystopia';
     const TokenASymbol = 'WETH';
     const tokenA = Tokens[Network.POLYGON][TokenASymbol];
     const TokenBSymbol = 'WMATIC';
@@ -76,7 +75,7 @@ describe('Dystopia', function () {
     it('getPoolIdentifiers and getPricesVolume', async function () {
       const dexHelper = new DummyDexHelper(Network.POLYGON);
       const blocknumber = await dexHelper.provider.getBlockNumber();
-      const dystopia = new DystopiaVolatile(Network.POLYGON, dexKey, dexHelper);
+      const dystopia = new Dystopia(Network.POLYGON, dexKey, dexHelper);
       const pools = await dystopia.getPoolIdentifiers(
         tokenA,
         tokenB,
@@ -109,7 +108,7 @@ describe('Dystopia', function () {
 
     it('getTopPoolsForToken', async function () {
       const dexHelper = new DummyDexHelper(Network.POLYGON);
-      const dystopia = new DystopiaVolatile(Network.POLYGON, dexKey, dexHelper);
+      const dystopia = new Dystopia(Network.POLYGON, dexKey, dexHelper);
 
       const poolLiquidity = await dystopia.getTopPoolsForToken(
         tokenA.address,
@@ -121,7 +120,7 @@ describe('Dystopia', function () {
     });
   });
   describe('Curve like stable pool', function () {
-    const dexKey = 'DystopiaStable';
+    const dexKey = 'Dystopia';
     const TokenASymbol = 'USDC';
     const tokenA = Tokens[Network.POLYGON][TokenASymbol];
     const TokenBSymbol = 'USDT';
@@ -130,11 +129,7 @@ describe('Dystopia', function () {
     it('getPoolIdentifiers and getPricesVolume', async function () {
       const dexHelper = new DummyDexHelper(Network.POLYGON);
       const blocknumber = await dexHelper.provider.getBlockNumber();
-      const dystopiaStable = new DystopiaStable(
-        Network.POLYGON,
-        dexKey,
-        dexHelper,
-      );
+      const dystopiaStable = new Dystopia(Network.POLYGON, dexKey, dexHelper);
       const pools = await dystopiaStable.getPoolIdentifiers(
         tokenA,
         tokenB,
@@ -167,11 +162,7 @@ describe('Dystopia', function () {
 
     it('getTopPoolsForToken', async function () {
       const dexHelper = new DummyDexHelper(Network.POLYGON);
-      const dystopiaStable = new DystopiaStable(
-        Network.POLYGON,
-        dexKey,
-        dexHelper,
-      );
+      const dystopiaStable = new Dystopia(Network.POLYGON, dexKey, dexHelper);
 
       const poolLiquidity = await dystopiaStable.getTopPoolsForToken(
         tokenA.address,
